requirements.txt
feedparser==6.0.11
pydantic==2.8.2
chromadb==0.4.24
fastmcp==1.0
Python-dotenv==1.0.1
numpy<2.0.0
